Refactor UI updaters into modular components
