+++ cannaclicker/src/app/ui.ts
@@ -2,7 +2,7 @@ import { createAudioManager } from './audio';
 import { t } from './i18n';
 import { evaluateAchievements, recalcDerivedValues } from './game';
 import type { GameState } from './state';
-import { initUI, type UIInitResult } from './ui';
+import { initUI, type UIInitResult } from './ui/bootstrap';
 
 
 interface UITicker {
